requirements.txt
pandas==1.3.4
numpy==1.21.2
psycopg2-binary==2.8.6